+++ src/baseConfig.js
@@ -1,4 +1,4 @@
-import authUtils from './authUtils';
+import {authUtils} from './authUtils';
 
 export class BaseConfig {
   configure(incomingConfig) {
@@ -11,42 +11,99 @@ export class BaseConfig {
 
   constructor() {
     this._current = {
-      endpoint: null, // Endpoint to use for auth requests (local only). Null uses HttpClient.
-      configureEndpoints: null, // Null to skip, or array of endpoints to patch.
-      client: null, // Used internally. The used Rest instance; set during configuration (see index.js)
-      httpInterceptor: true,
-      loginOnSignup: true,
-      baseUrl: null,
-      loginRedirect: '/#customer',
-      logoutRedirect: '/',
-      signupRedirect: '/login',
-      loginUrl: '/auth/login',
-      signupUrl: '/auth/signup',
-      profileUrl: '/auth/me',
+      // Used internally. The used Rest instance; set during configuration (see index.js)
+      client: null,
+
+      // If using aurelia-api:
+      // =====================
+
+      // This is the endpoint used for any requests made in relation to authentication (login, logout, etc.)
+      endpoint: null,
+      // When authenticated, these endpoints will have the token added to the header of any requests (for authorization)
+      configureEndpoints: null,
+
+
+      // SPA related options
+      // ===================
+
+      // The SPA url to which the user is redirected after a successful login
+      loginRedirect: '#/customer',
+      // The SPA url to which the user is redirected after a successful logout
+      logoutRedirect: '#/',
+      // The SPA route used when an unauthenticated user tries to access an SPA page that requires authentication
       loginRoute: '/login',
-      signupRoute: '/signup',
-      postContentType: 'json', //options form|json
+      // Whether or not an authentication token is provided in the response to a successful signup
+      loginOnSignup: true,
+      // If loginOnSignup == false: The SPA url to which the user is redirected after a successful signup (else loginRedirect is used)
+      signupRedirect: '#/login',
+
+
+      // API related options
+      // ===================
+
+      // The base url used for all authentication related requests, including
+      // provider.url bellow. This appends to the httpClient/endpoint base url,
+      //  it does not override it.
+      baseUrl: '/auth',
+      // The API endpoint to which login requests are sent
+      loginUrl: '/login',
+      // The API endpoint to which signup requests are sent
+      signupUrl: '/signup',
+      // The API endpoint used in profile requests (inc. `find/get` and `update`)
+      profileUrl: '/me',
+      // The API endpoint used with oAuth to unlink authentication
+      unlinkUrl: '/unlink/',
+    	// The HTTP method used for 'unlink' requests (Options: 'get' or 'post')
       useRefreshToken: false,
       autoUpdateToken: true,
       refreshTokenRoot: false,
       refreshTokenName: 'refresh_token',
       refreshTokenPrefix: 'aurelia',
       clientId: false,
-      tokenRoot: false,
-      tokenName: 'token',
-      tokenPrefix: 'aurelia',
-      responseTokenProp: 'access_token',
-      unlinkUrl: '/auth/unlink/',
       unlinkMethod: 'get',
+
+
+      // Token Related options
+      // =====================
+
+      // The header property used to contain the authToken in the header of API requests that require authentication
       authHeader: 'Authorization',
+      // The token name used in the header of API requests that require authentication
       authToken: 'Bearer',
+      // The the property from which to get the authentication token after a successful login or signup
+      responseTokenProp: 'access_token',
+
+      // If the property defined by `responseTokenProp` is an object:
+      // ------------------------------------------------------------
+
+      //This is the property from which to get the token `{ "responseTokenProp": { "tokenName" : '...' } }`
+      tokenName: 'token',
+      // This allows the token to be a further object deeper `{ "responseTokenProp": { "tokenRoot" : { "tokenName" : '...' } } }`
+      tokenRoot: false,
+
+
+      // Miscellaneous Options
+      // =====================
+
+      // Whether to enable the fetch interceptor which automatically adds the authentication headers
+      // (or not... e.g. if using a session based API or you want to override the default behaviour)
+      httpInterceptor: true,
+      // For OAuth only: Tell the API whether or not to include token cookies in the response (for session based APIs)
       withCredentials: true,
+      // Controls how the popup is shown for different devices (Options: 'browser' or 'mobile')
       platform: 'browser',
+      // Determines the `window` property name upon which aurelia-auth data is stored (Default: `window.localStorage`)
       storage: 'localStorage',
+      // Prepended to the `tokenName` when kept in storage (nothing to do with)
+      tokenPrefix: 'aurelia',
+
+
+      //OAuth provider specific related configuration
+      // ============================================
       providers: {
         google: {
           name: 'google',
-          url: '/auth/google',
+          url: '/google',
           authorizationEndpoint: 'https://accounts.google.com/o/oauth2/auth',
           redirectUri: window.location.origin || window.location.protocol + '//' + window.location.host,
           scope: ['profile', 'email'],
@@ -64,7 +121,7 @@ export class BaseConfig {
         },
         facebook: {
           name: 'facebook',
-          url: '/auth/facebook',
+          url: '/facebook',
           authorizationEndpoint: 'https://www.facebook.com/v2.3/dialog/oauth',
           redirectUri: window.location.origin + '/' || window.location.protocol + '//' + window.location.host + '/',
           scope: ['email'],
@@ -82,7 +139,7 @@ export class BaseConfig {
         },
         linkedin: {
           name: 'linkedin',
-          url: '/auth/linkedin',
+          url: '/linkedin',
           authorizationEndpoint: 'https://www.linkedin.com/uas/oauth2/authorization',
           redirectUri: window.location.origin || window.location.protocol + '//' + window.location.host,
           requiredUrlParams: ['state'],
@@ -97,7 +154,7 @@ export class BaseConfig {
         },
         github: {
           name: 'github',
-          url: '/auth/github',
+          url: '/github',
           authorizationEndpoint: 'https://github.com/login/oauth/authorize',
           redirectUri: window.location.origin || window.location.protocol + '//' + window.location.host,
           optionalUrlParams: ['scope'],
@@ -111,7 +168,7 @@ export class BaseConfig {
         },
         yahoo: {
           name: 'yahoo',
-          url: '/auth/yahoo',
+          url: '/yahoo',
           authorizationEndpoint: 'https://api.login.yahoo.com/oauth2/request_auth',
           redirectUri: window.location.origin || window.location.protocol + '//' + window.location.host,
           scope: [],
@@ -124,7 +181,7 @@ export class BaseConfig {
         },
         twitter: {
           name: 'twitter',
-          url: '/auth/twitter',
+          url: '/twitter',
           authorizationEndpoint: 'https://api.twitter.com/oauth/authenticate',
           type: '1.0',
           popupOptions: {
@@ -149,7 +206,7 @@ export class BaseConfig {
         },
         instagram: {
           name: 'instagram',
-          url: '/auth/instagram',
+          url: '/instagram',
           authorizationEndpoint: 'https://api.instagram.com/oauth/authorize',
           redirectUri: window.location.origin || window.location.protocol + '//' + window.location.host,
           requiredUrlParams: ['scope'],